+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/fluent/SeleniumCheckSettings.java
@@ -440,6 +440,18 @@ public class SeleniumCheckSettings extends CheckSettings implements ISeleniumChe
         return this.ignoreDisplacements(true);
     }
 
+    public SeleniumCheckSettings enablePatterns(boolean enablePatterns) {
+        SeleniumCheckSettings clone = this.clone();
+        clone.enablePatterns = enablePatterns;
+        return clone;
+    }
+
+    public SeleniumCheckSettings enablePatterns() {
+        SeleniumCheckSettings clone = this.clone();
+        clone.enablePatterns = true;
+        return clone;
+    }
+
     public SeleniumCheckSettings statusBarExists() {
         SeleniumCheckSettings clone = this.clone();
         clone.statusBarExists = true;