+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/ios/IOSVisualLocatorTest.java
@@ -1,11 +1,13 @@
-package com.applitools.eyes.appium;
+package com.applitools.eyes.appium.ios;
 
 import com.applitools.eyes.Location;
+import com.applitools.eyes.LogHandler;
 import com.applitools.eyes.Region;
 import com.applitools.eyes.StdoutLogHandler;
+import com.applitools.eyes.appium.Eyes;
 import com.applitools.eyes.locators.VisualLocator;
 import io.appium.java_client.TouchAction;
-import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.ios.IOSDriver;
 import io.appium.java_client.touch.WaitOptions;
 import io.appium.java_client.touch.offset.PointOption;
 import org.openqa.selenium.remote.DesiredCapabilities;
@@ -15,37 +17,33 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 
-public class AndroidVisualLocatorsTest {
+public class IOSVisualLocatorTest {
 
     public static void main(String[] args) throws Exception {
-
         DesiredCapabilities capabilities = new DesiredCapabilities();
-        // IMPORTANT: This assumes an emulator running locally. Make sure the settings match the device you're running.
-        capabilities.setCapability("platformName", "Android");
-        capabilities.setCapability("deviceName", "Google Nexus 6");
-        capabilities.setCapability("platformVersion", "7.1.1");
 
-        capabilities.setCapability("app", "https://applitools.bintray.com/Examples/app-android.apk");
-        capabilities.setCapability("automationName", "UiAutomator2");
+        capabilities.setCapability("platformName", "iOS");
+        capabilities.setCapability("deviceName", "iPhone Xr");
+        capabilities.setCapability("platformVersion", "12.4");
+        capabilities.setCapability("app", "/Users/alexanderkachechka/Downloads/app-ios.app");
+        capabilities.setCapability("automationName", "XCUITest");
+        capabilities.setCapability("fullReset", true);
         capabilities.setCapability("newCommandTimeout", 300);
 
-        AndroidDriver driver = new AndroidDriver(new URL("http://127.0.0.1:4723/wd/hub"), capabilities);
-
-        driver.manage().timeouts().implicitlyWait(10_000, TimeUnit.MILLISECONDS);
+        IOSDriver driver = new IOSDriver<>(new URL("http://127.0.0.1:4723/wd/hub"), capabilities);
 
         Eyes eyes = new Eyes();
 
-        eyes.setLogHandler(new StdoutLogHandler(true));
+        LogHandler logHandler = new StdoutLogHandler(true);
+        eyes.setLogHandler(logHandler);
         eyes.setForceFullPageScreenshot(false);
+        eyes.setScrollToRegion(false);
         eyes.setMatchTimeout(1000);
-        eyes.setSaveDebugScreenshots(false);
-
         try {
-            eyes.open(driver, "Android Test App", "Test Visual Locators");
+            eyes.open(driver, "iOS Test App", "Test Visual Locators");
 
-            eyes.checkWindow("Launch screen test");
+            eyes.checkWindow("Launch screen");
 
             Map<String, List<Region>> locators = eyes.locate(VisualLocator.name("list_view_locator").name("scroll_view_locator"));
             System.out.println("Received locators" + locators);
@@ -72,10 +70,6 @@ public class AndroidVisualLocatorsTest {
 
                 eyes.checkWindow("ListView screen");
             }
-
-            locators = eyes.locate(VisualLocator.name("header_locator"));
-            System.out.println("Received locators" + locators);
-
             eyes.close();
         } finally {
             driver.quit();