+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ### Updated 
 - Ignore displacements and enable patterns to Configuration. [Trello 844](https://trello.com/c/CvOpWbfU)
+### Added
+- stitchingAdjustment parameter to Configuration class. It will be used to avoid shadows on stitching parts. [Trello 1419](https://trello.com/c/IIIKosP8)
 
 ## [4.12.1]
 ### Added